trying to make it dynamic
