+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import 'widgets/stepper_widget.dart';
+
 void main() {
   runApp(const MyApp());
 }
@@ -31,48 +33,6 @@ class MyHomePage extends StatelessWidget {
     return Scaffold(
       appBar: AppBar(),
       body: const StepperWidget(),
-    );
-  }
-}
-
-class StepperWidget extends StatefulWidget {
-  const StepperWidget({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  State<StepperWidget> createState() => _StepperWidgetState();
-}
-
-class _StepperWidgetState extends State<StepperWidget> {
-  int _index = 0;
-  @override
-  Widget build(BuildContext context) {
-    return Stepper(
-      currentStep: _index,
-      onStepContinue: () {
-        if (_index >= 3) return;
-        setState(() {
-          _index++;
-        });
-      },
-      onStepCancel: () {
-        if (_index <= 0) return;
-        setState(() {
-          _index--;
-        });
-      },
-      onStepTapped: ((value) {
-        setState(() {
-          _index = value;
-        });
-      }),
-      steps: const <Step>[
-        Step(title: Text("User Details"), content: Text("test")),
-        Step(title: Text("Address"), content: Text("test")),
-        Step(title: Text("Profession Details"), content: Text("test")),
-        Step(title: Text("Confirm"), content: Text("test")),
-      ],
     );
   }
 }
